cluster/validation: move authentik blueprint code into its own module
Lift the authentik blueprint parsing (the `!Find`/`!KeyOf` SafeLoader + Pydantic
entry models) and both authentik checks — check_proxy_provider_outpost_assignment
and the pre-existing check_blueprint_completeness — out of the grab-bag checks.py
into authentik_blueprints.py, matching the per-domain check-module pattern
(image_automation.py, terraform_backends.py). No behaviour change.

BAZEL_TEST_INVOCATIONS=buildbuddy:f9c38620-4703-41ff-9ec9-7b0b0631301b
