+++ cluster/validation/validate_all.py
@@ -31,12 +31,14 @@ import asyncio
 import os
 from pathlib import Path
 
-from cluster.validation.checks import (
+from cluster.validation.authentik_blueprints import (
     check_blueprint_completeness,
+    check_proxy_provider_outpost_assignment,
+)
+from cluster.validation.checks import (
     check_duplicate_external_secrets,
     check_goldilocks_explicit_decision,
     check_goldilocks_namespace_labels,
-    check_proxy_provider_outpost_assignment,
     find_orphaned_files,
 )
 from cluster.validation.cluster import parse_cluster
